requirements.txt
pyodbc
dotenv
azure-search-documents>=11.4.0b6
openai>=1.0.0
pathlib
uuid
streamlit
pandas